backend/requirements.txt
alembic==0.8.10
bleach==2.0.0
blinker==1.4
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==6.7
dominate==2.7.0
firebase-admin==6.0.0
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==3.2.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
gcloud==0.18.3
google-api-core==2.10.1
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
html5lib==1.0b10
httplib2==0.20.4
idna==3.4
itsdangerous==0.24
Jinja2==2.9.5
jws==0.1.3
Mako==1.0.6
Markdown==2.6.8
MarkupSafe==0.23
msgpack==1.0.4
oauth2client==4.1.3
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.21.0
python-editor==1.0.3
python-jwt==2.0.1
requests==2.28.1
requests-toolbelt==0.10.0
rsa==4.9
six==1.10.0
SQLAlchemy==1.1.5
swagger-gen==0.1.2
uritemplate==4.1.1
urllib3==1.26.12
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.11.15
WTForms==2.1